+++ .github/ISSUE_TEMPLATE/bug-report.yml
@@ -7,7 +7,7 @@ body:
 - type: markdown
   attributes:
     value: >
-      **Thank you for wanting to report a bug in Arx-AST!**
+      **Thank you for wanting to report a bug in AstX!**
 
 
       ⚠
@@ -15,7 +15,7 @@ body:
       GitHub][issue search].
 
 
-      [issue search]: https://github.com/Arx-AST/issues?q=is%3Aopen+is%3Aissue+label%3Abug
+      [issue search]: https://github.com/AstX/issues?q=is%3Aopen+is%3Aissue+label%3Abug
 
 - type: textarea
   attributes:
@@ -124,14 +124,14 @@ body:
 
 - type: textarea
   attributes:
-    # label: Arx-AST version
+    # label: AstX version
     # description: >-
-     # Paste verbatim output from `Arx-AST --version` below, under the prompt line.
+     # Paste verbatim output from `AstX --version` below, under the prompt line.
      # Don't wrap it with triple backticks &mdash; your whole input will be
      # turned into a code snippet automatically.
     render: console
     placeholder: |
-      python -m pip show Arx-AST
+      python -m pip show AstX
   validations:
     required: true
 
@@ -139,9 +139,9 @@ body:
   attributes:
     label: Code of Conduct
     description: |
-      Read the [`Arx-AST` Code of Conduct][CoC] first.
+      Read the [`AstX` Code of Conduct][CoC] first.
 
-      [CoC]: https://github.com/arx-ast/coc/CODE_OF_CONDUCT.md
+      [CoC]: https://github.com/astx/coc/CODE_OF_CONDUCT.md
     options:
     - label: I agree to follow the Code of Conduct
       required: true
