+++ .github/ISSUE_TEMPLATE/config.yml
@@ -2,14 +2,14 @@
 blank_issues_enabled: true
 contact_links:
 - name: 🙋  Usage Questions
-  url: https://github.com/arx-ast/discussions
+  url: https://github.com/astx/discussions
   about: |
-    Use Arx-AST's GitHub Discussions to ask "How do I do X with Cookiecutter?".
+    Use AstX's GitHub Discussions to ask "How do I do X with Cookiecutter?".
 - name: 📖  Tutorial
-  url: https://github.com/arx-ast
+  url: https://github.com/astx
   about: |
-    The Arx-AST tutorial is continually updated and provides an in depth walkthrough
+    The AstX tutorial is continually updated and provides an in depth walkthrough
     of how to use the latest release of Cookiecutter.
-- name: 📝  Arx-AST Code of Conduct
-  url: https://github.com/arx-ast/coc/CODE_OF_CONDUCT.md
+- name: 📝  AstX Code of Conduct
+  url: https://github.com/astx/coc/CODE_OF_CONDUCT.md
   about: Expectations for how people will interact with each other on Cookiecutter's GitHub.
