+++ .github/ISSUE_TEMPLATE/documentation-report.yml
@@ -7,7 +7,7 @@ body:
 - type: markdown
   attributes:
     value: >
-      **Thank you for wanting to report a problem with Arx-AST's documentation!**
+      **Thank you for wanting to report a problem with AstX's documentation!**
 
 
       ⚠
@@ -15,7 +15,7 @@ body:
       GitHub][issue search].
 
 
-      [issue search]: https://github.com/Arx-AST/issues?q=is%3Aopen+is%3Aissue+label%3Adocs
+      [issue search]: https://github.com/AstX/issues?q=is%3Aopen+is%3Aissue+label%3Adocs
 
 - type: textarea
   attributes:
@@ -25,7 +25,7 @@ body:
       If there are external references that are related please link them here
       as well.
     placeholder: >-
-      I was reading the Arx-AST documentation for Arx-AST version X and I'm having
+      I was reading the AstX documentation for AstX version X and I'm having
       problems understanding Y.
       It would be very helpful if that got rephrased as Z.
   validations:
@@ -43,9 +43,9 @@ body:
   attributes:
     label: Code of Conduct
     description: |
-      Read the [`Arx-AST` Code of Conduct][CoC] first.
+      Read the [`AstX` Code of Conduct][CoC] first.
 
-      [CoC]: https://github.com/Arx-AST/blob/main/CODE_OF_CONDUCT.md
+      [CoC]: https://github.com/AstX/blob/main/CODE_OF_CONDUCT.md
     options:
     - label: I agree to follow the Code of Conduct
       required: true
