+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
       name: mypy
       entry: mypy
       language: system
-      files: "src/arxast"
+      files: "src/astx"
       pass_filenames: true
       types:
         - python
@@ -51,7 +51,7 @@ repos:
       name: vulture
       entry: vulture --min-confidence 80
       language: system
-      files: "src/arxast"
+      files: "src/astx"
       description: Find unused Python code.
       pass_filenames: true
       types:
@@ -61,7 +61,7 @@ repos:
       name: mccabe
       entry: python -m mccabe --min 10
       language: system
-      files: "src/arxast"
+      files: "src/astx"
       pass_filenames: true
       types:
         - python
