+++ docs/api/references.md
@@ -1,3 +1,3 @@
 # Api references
 
-::: arxast
+::: astx
