+++ docs/installation.md
@@ -2,14 +2,14 @@
 
 ## Stable release
 
-To install Arx-AST, run this command in your
+To install AstX, run this command in your
 terminal:
 
 ```bash
-$ pip install arx-ast
+$ pip install astx
 ```
 
-This is the preferred method to install Arx-AST,
+This is the preferred method to install AstX,
 as it will always install the most recent stable release.
 
 If you don't have [pip](https://pip.pypa.io) installed, this
@@ -18,20 +18,20 @@ can guide you through the process.
 
 ## From sources
 
-The sources for Arx-AST can be downloaded from
-the [Github repo](https://github.com/arxlang/arx-ast).
+The sources for AstX can be downloaded from
+the [Github repo](https://github.com/arxlang/astx).
 
 You can either clone the public repository:
 
 ```bash
-$ git clone https://github.com/arxlang/arx-ast
+$ git clone https://github.com/arxlang/astx
 ```
 
 Or download the
-[tarball](https://github.com/arxlang/arx-ast/tarball/main):
+[tarball](https://github.com/arxlang/astx/tarball/main):
 
 ```bash
-$ curl -OJL https://github.com/arxlang/arx-ast/tarball/main
+$ curl -OJL https://github.com/arxlang/astx/tarball/main
 ```
 
 Once you have a copy of the source, you can install it with:
