+++ src/astx/blocks.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import List
 
-from arxast.base import AST, Expr, ASTKind
+from astx.base import AST, Expr, ASTKind
 
 
 class Block(AST):
